+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/EngineGetPayloadResultV3.java
@@ -27,11 +27,12 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import org.apache.tuweni.bytes.Bytes32;
 
-@JsonPropertyOrder({"executionPayload", "blockValue", "blobsBundle"})
+@JsonPropertyOrder({"executionPayload", "blockValue", "blobsBundle", "shouldOverrideBuilder"})
 public class EngineGetPayloadResultV3 {
   protected final PayloadResult executionPayload;
   private final String blockValue;
   private final BlobsBundleV1 blobsBundle;
+  private final boolean shouldOverrideBuilder;
 
   public EngineGetPayloadResultV3(
       final BlockHeader header,
@@ -42,6 +43,7 @@ public class EngineGetPayloadResultV3 {
     this.executionPayload = new PayloadResult(header, transactions, withdrawals);
     this.blockValue = blockValue;
     this.blobsBundle = blobsBundle;
+    this.shouldOverrideBuilder = false;
   }
 
   @JsonGetter(value = "executionPayload")
@@ -59,6 +61,11 @@ public class EngineGetPayloadResultV3 {
     return blobsBundle;
   }
 
+  @JsonGetter(value = "shouldOverrideBuilder")
+  public boolean shouldOverrideBuilder() {
+    return shouldOverrideBuilder;
+  }
+
   public static class PayloadResult {
 
     protected final String blockHash;
@@ -74,10 +81,9 @@ public class EngineGetPayloadResultV3 {
     private final String timestamp;
     private final String extraData;
     private final String baseFeePerGas;
-
     private final String excessBlobGas;
-
     private final String blobGasUsed;
+    private final String parentBeaconBlockRoot;
 
     protected final List<String> transactions;
     private final List<WithdrawalParameter> withdrawals;
@@ -111,6 +117,8 @@ public class EngineGetPayloadResultV3 {
       this.blobGasUsed = header.getBlobGasUsed().map(Quantity::create).orElse(Quantity.HEX_ZERO);
       this.excessBlobGas =
           header.getExcessBlobGas().map(Quantity::create).orElse(Quantity.HEX_ZERO);
+      this.parentBeaconBlockRoot =
+          header.getParentBeaconBlockRoot().map(Bytes32::toHexString).orElse(null);
     }
 
     @JsonGetter(value = "blockNumber")
@@ -197,6 +205,11 @@ public class EngineGetPayloadResultV3 {
     @JsonGetter(value = "blobGasUsed")
     public String getBlobGasUseds() {
       return blobGasUsed;
+    }
+
+    @JsonGetter(value = "parentBeaconBlockRoot")
+    public String getParentBeaconBlockRoot() {
+      return parentBeaconBlockRoot;
     }
   }
 }