+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/chain/GenesisState.java
@@ -53,6 +53,7 @@ import java.util.stream.Stream;
 
 import com.google.common.base.MoreObjects;
 import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.units.bigints.UInt256;
 
 public final class GenesisState {
@@ -170,6 +171,8 @@ public final class GenesisState {
         .withdrawalsRoot(isShanghaiAtGenesis(genesis) ? Hash.EMPTY_TRIE_HASH : null)
         .blobGasUsed(isCancunAtGenesis(genesis) ? parseBlobGasUsed(genesis) : null)
         .excessBlobGas(isCancunAtGenesis(genesis) ? parseExcessBlobGas(genesis) : null)
+        .parentBeaconBlockRoot(
+            (isCancunAtGenesis(genesis) ? parseParentBeaconBlockRoot(genesis) : null))
         .depositsRoot(isExperimentalEipsTimeAtGenesis(genesis) ? Hash.EMPTY_TRIE_HASH : null)
         .buildBlockHeader();
   }
@@ -230,6 +233,11 @@ public final class GenesisState {
         withNiceErrorMessage(
             "excessBlobGas", genesis.getExcessBlobGas(), GenesisState::parseUnsignedLong);
     return BlobGas.of(excessBlobGas);
+  }
+
+  private static Bytes32 parseParentBeaconBlockRoot(final GenesisConfigFile genesis) {
+    return withNiceErrorMessage(
+        "parentBeaconBlockRoot", genesis.getParentBeaconBlockRoot(), Bytes32::fromHexString);
   }
 
   private static long parseUnsignedLong(final String value) {