+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/messages/MessageWrapperTest.java
@@ -392,6 +392,7 @@ public class MessageWrapperTest {
           null,
           null,
           null,
+          null,
           new MainnetBlockHeaderFunctions());
     }
   }