+++ ethereum/referencetests/src/main/java/org/hyperledger/besu/ethereum/referencetests/StateTestVersionedTransaction.java
@@ -18,11 +18,11 @@ package org.hyperledger.besu.ethereum.referencetests;
 import org.hyperledger.besu.crypto.KeyPair;
 import org.hyperledger.besu.crypto.SignatureAlgorithm;
 import org.hyperledger.besu.crypto.SignatureAlgorithmFactory;
+import org.hyperledger.besu.datatypes.AccessListEntry;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.VersionedHash;
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.ethereum.core.Transaction;
-import org.hyperledger.besu.evm.AccessListEntry;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
@@ -73,7 +73,8 @@ public class StateTestVersionedTransaction {
   private final List<Bytes> payloads;
   private final Optional<List<List<AccessListEntry>>> maybeAccessLists;
   private final Wei maxFeePerBlobGas;
-  private final List<VersionedHash> blobVersionedHashes;
+  // String instead of VersionedHash because reference tests intentionally use bad hashes.
+  private final List<String> blobVersionedHashes;
 
   /**
    * Constructor for populating a mock transaction with json data.
@@ -103,7 +104,8 @@ public class StateTestVersionedTransaction {
       @JsonDeserialize(using = StateTestAccessListDeserializer.class) @JsonProperty("accessLists")
           final List<List<AccessListEntry>> maybeAccessLists,
       @JsonProperty("maxFeePerBlobGas") final String maxFeePerBlobGas,
-      @JsonProperty("blobVersionedHashes") final List<VersionedHash> blobVersionedHashes) {
+      @JsonProperty("maxFeePerDataGas") final String maxFeePerDataGas,
+      @JsonProperty("blobVersionedHashes") final List<String> blobVersionedHashes) {
 
     this.nonce = Bytes.fromHexStringLenient(nonce).toLong();
     this.gasPrice = Optional.ofNullable(gasPrice).map(Wei::fromHexString).orElse(null);
@@ -122,7 +124,9 @@ public class StateTestVersionedTransaction {
     this.payloads = parseArray(data, Bytes::fromHexString);
     this.maybeAccessLists = Optional.ofNullable(maybeAccessLists);
     this.maxFeePerBlobGas =
-        Optional.ofNullable(maxFeePerBlobGas).map(Wei::fromHexString).orElse(null);
+        Optional.ofNullable(maxFeePerBlobGas == null ? maxFeePerDataGas : maxFeePerBlobGas)
+            .map(Wei::fromHexString)
+            .orElse(null);
     this.blobVersionedHashes = blobVersionedHashes;
   }
 
@@ -161,7 +165,15 @@ public class StateTestVersionedTransaction {
     maybeAccessLists.ifPresent(
         accessLists -> transactionBuilder.accessList(accessLists.get(indexes.data)));
     Optional.ofNullable(maxFeePerBlobGas).ifPresent(transactionBuilder::maxFeePerBlobGas);
-    transactionBuilder.versionedHashes(blobVersionedHashes);
+    try {
+      transactionBuilder.versionedHashes(
+          blobVersionedHashes == null
+              ? null
+              : blobVersionedHashes.stream().map(VersionedHash::fromHexString).toList());
+    } catch (IllegalArgumentException iae) {
+      // versioned hash string was bad, so this is an invalid transaciton
+      return null;
+    }
 
     transactionBuilder.guessType();
     if (transactionBuilder.getTransactionType().requiresChainId()) {