+++ evm/src/main/java/org/hyperledger/besu/evm/frame/MessageFrame.java
@@ -994,6 +994,17 @@ public class MessageFrame {
   }
 
   /**
+   * Returns whether an address has been warmed up. Is deliberately publicly exposed for access from
+   * tracers.
+   *
+   * @param address the address context
+   * @return whether the address has been warmed up
+   */
+  public boolean isAddressWarm(final Address address) {
+    return txValues.warmedUpAddresses().contains(address);
+  }
+
+  /**
    * "Warms up" the storage slot as per EIP-2929
    *
    * @param address the address whose storage is being warmed up
@@ -1002,6 +1013,18 @@ public class MessageFrame {
    */
   public boolean warmUpStorage(final Address address, final Bytes32 slot) {
     return txValues.warmedUpStorage().put(address, slot, Boolean.TRUE) != null;
+  }
+
+  /**
+   * Returns whether an address' slot is warmed up. Is deliberately publicly exposed for access from
+   * trace
+   *
+   * @param address the address context
+   * @param slot the slot to query
+   * @return whether the address/slot couple is warmed up
+   */
+  public boolean isStorageWarm(final Address address, final Bytes32 slot) {
+    return this.txValues.warmedUpStorage().contains(address, slot);
   }
 
   /**